+++ index.html
@@ -248,7 +248,7 @@
     </div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 
 </html>
